+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **Update:** Fixed filter li display blank space bug ([#431](https://github.com/wenzhixin/multiple-select/pull/431)).
 - **Update:** Fixed container outer width bug ([#431](https://github.com/wenzhixin/multiple-select/pull/431)).
 - **Update:** Updated `addTitle` option to `displayTitle` option ([#431](https://github.com/wenzhixin/multiple-select/pull/431)).
+- **Update:** Fixed filter with no select options display bug ([#431](https://github.com/wenzhixin/multiple-select/pull/431)).
 
 ### 1.3.0